refactor(glyph-picker): extract inner panel for Studio Look tab reuse
The existing modal keeps its outer chrome (backdrop, header preview,
footer) and delegates its body to a new FamiliarGlyphPickerPanel.
The Studio drawer's Look tab will embed the panel inline. No
behavior change to the existing modal trigger.
